backend/pyproject.toml
[tool.poetry]
name = "pythocron"
version = "0.1.0"
description = "pythocron"
authors = ["Oskar Korczak <oskar.korczak@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.0"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
python-crontab = "^2.6.0"
base58 = "^2.1.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
